+++ xcube/core/gen2/local/describer.py
@@ -73,7 +73,7 @@ class DatasetsDescriber:
                                               data_type=DATASET_TYPE)
         except DataStoreError as dse:
             raise CubeGeneratorError(f'{dse}',
-                                     status_code=400)
+                                     status_code=400) from dse
         if not isinstance(descriptor, DatasetDescriptor):
             raise RuntimeError(f'internal error: data store '
                                f'"{input_config.store_id}": '